Gercek konum ve SLAM izi: veri hatti, hassasiyet ve cizim duzeltmeleri
Yer istasyonu artik gercek konumu (ground truth) mavi, SLAM kestirimini
turuncu olarak biriktirerek cizer; buton yok, her poz mesajinda uzar.

Duzeltmeler:
- pose/slamPose lat/lon float -> double. float32 39 enleminde ~1.1 m
  kuantizasyon demekti ve desimetre altindaki SLAM sapmasini tamamen
  siliyordu.
- Iz min-adim kontrolu artik son eklenen vertex'e gore yapiliyor. Onceki
  halde atlanan noktada referans da guncellendigi icin kucuk adimlar hic
  birikmiyor ve yogun ornekli akislarda iz tek noktada takili kaliyordu.
- Iz varsayilan olarak duz ciziliyor; arazi yukseklik sorgusu tile'lar
  yuklenirken degistigi icin cizgi dikeyde sicriyordu.
- Cizgi kalinligi metre cinsinden verilip harita olcegine cevriliyor ve
  zoom degisince guncelleniyor. SLAM izi daha ince ve ustte cizilir,
  boylece izler bindiginde gercek konum kaybolmaz.
- Tampon dolunca tek tek degil toplu kirpma (her fix'te tam rebuild yoktu).

Eklenenler:
- DigitalTwinTrajectoryDatasetPlayer: EuRoC/TUM yorunge dosyalarini okur,
  zaman damgasina gore eslestirir, benzerlik donusumuyle hizalar ve
  kopruye besler (harici surec gerekmez).
- Tools/SlamFeeder: ORB-SLAM3 ciktisi + ground truth icin Python besleyici
  (Sim(3)/SE(3) hizalama, ATE/RMSE raporu, UDP 19090), dinleyici, veri
  indirici ve 16 test. ORB-SLAM3 ciktisi yokken --simulate-slam ile ayni
  bicimde kestirim uretilir.
- TrajectorySmokeTest: Play modunda izlerin gercekten cizildigini olcen
  editor testi (rapor + kanit karesi).
- Test verisi: EuRoC MH_01_easy ground truth (CC BY 3.0, atif data/README).

Dogrulama: Unity 2022.3.62f3 Play modu, MH01 -> 140/143 vertex, sapma ort
0.176 m, RMSE 0.207 m, hizalama olcegi 1.2683 (Python Umeyama 1.2689 ile
tutuyor), 0 konsol hatasi.

+++ Assets/Scripts/DigitalTwin/Contracts/TwinMissionContracts.cs
@@ -5,8 +5,10 @@ namespace GroundStation.DigitalTwin
     [Serializable]
     public class TwinSlamPoseBlock
     {
-        public float latitude;
-        public float longitude;
+        // TwinPoseBlock ile ayni gerekce: SLAM kestirimi metre alti farklari tasir,
+        // float lat/lon bu farki sifirlar. Bkz. DigitalTwinJsonSchema.TwinPoseBlock.
+        public double latitude;
+        public double longitude;
         public float altitudeM;
         public float yawDeg;
         public float pitchDeg;
